move wrongly placed alpak functions from `alpaka::onHost` to `alpaka`
Move `getNumElemPerThread()` , `getArchSimdWidth()`,
  `getNumPipelines()`, `getDeviceKind()` and `getApi()'from host
namespace to alpaka
