+++ example/tutorial/src/04_views.cpp
@@ -43,8 +43,8 @@ int example(auto const devSpec)
     {
         // allocate a buffer of floats in global device memory
         auto device_buffer = alpaka::onHost::alloc<float>(device, size);
-        std::cout << "memory buffer on " << alpaka::onHost::getStaticName(alpaka::onHost::getApi(device_buffer))
-                  << " at " << std::data(device_buffer) << "\n\n";
+        std::cout << "memory buffer on " << alpaka::onHost::getStaticName(alpaka::getApi(device_buffer)) << " at "
+                  << std::data(device_buffer) << "\n\n";
 
         // set the device memory to all zeros (byte-wise, not element-wise)
         alpaka::onHost::memset(queue, device_buffer, 0x00);
