+++ include/alpaka/alpaka.hpp
@@ -18,6 +18,7 @@
 #include "alpaka/core/Utility.hpp"
 #include "alpaka/core/common.hpp"
 #include "alpaka/core/config.hpp"
+#include "alpaka/interface.hpp"
 #include "alpaka/internal.hpp"
 #include "alpaka/math.hpp"
 #include "alpaka/math/constants.hpp"
