+++ include/alpaka/api/unifiedCudaHip/Device.hpp
@@ -122,7 +122,7 @@ namespace alpaka::internal
     {
         inline constexpr auto operator()(auto&& device) const
         {
-            return onHost::getApi(device.m_platform);
+            return getApi(device.m_platform);
         }
     };
 } // namespace alpaka::internal
